+++ console/src/app/pages/orgs/password-age-policy/password-age-policy.component.html
@@ -7,8 +7,7 @@
         </button>
     </ng-template>
 
-
-    <div class="content" *ngIf="policyType === PolicyComponentType?.AGE && ageData">
+    <div class="content" *ngIf="ageData">
         <mat-form-field class="description-formfield" appearance="outline">
             <mat-label>{{ 'ORG.POLICY.DATA.DESCRIPTION' | translate }}</mat-label>
             <input matInput name="description" ngDefaultControl [(ngModel)]="ageData.description" required />
