+++ console/src/app/pages/orgs/password-iam-policy/password-iam-policy.component.scss
@@ -1,0 +1,44 @@
+
+button {
+  border-radius: .5rem;
+}
+
+.content {
+  padding-top: 1rem;
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+
+  .row {
+    display: flex;
+    align-items: center;
+    padding: .5rem 0;
+
+    .left-desc {
+      color: #8795a1;
+      font-size: .9rem;
+    }
+
+    .fill-space {
+      flex: 1;
+    }
+
+    .length-wrapper {
+      display: flex;
+      align-items: center;
+    }
+  }
+}
+
+.btn-container {
+  display: flex;
+  justify-content: flex-end;
+  width: 100%;
+
+  button {
+    margin-top: 3rem;
+    display: block;
+    padding: .5rem 4rem;
+    border-radius: .5rem;
+  }
+}
