--- console/src/app/pages/orgs/password-policy/password-policy.component.html
@@ -1,136 +1,0 @@
-<app-detail-layout [backRouterLink]="[ '/org']" [title]="title ? (title | translate) : ''"
-    [description]="desc ? (desc | translate) : ''">
-    <ng-template appHasRole [appHasRole]="['iam.policy.write']">
-        <button matTooltip="{{'ORG.POLICY.DELETE' | translate}}" color="warn" (click)="deletePolicy()"
-            mat-stroked-button>
-            {{'ORG.POLICY.DELETE' | translate}}
-        </button>
-    </ng-template>
-
-    <div class="content" *ngIf="policyType === PolicyComponentType?.LOCKOUT && lockoutData">
-        <mat-form-field class="description-formfield" appearance="outline">
-            <mat-label>{{ 'ORG.POLICY.DATA.DESCRIPTION' | translate }}</mat-label>
-            <input matInput name="description" ngDefaultControl [(ngModel)]="lockoutData.description" required />
-        </mat-form-field>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.MAXATTEMPTS' | translate}}</span>
-            <span class="fill-space"></span>
-            <div class="length-wrapper">
-                <button mat-icon-button (click)="incrementMaxAttempts()">
-                    <mat-icon>add</mat-icon>
-                </button>
-                <span>{{lockoutData?.maxAttempts}}</span>
-                <button mat-icon-button (click)="decrementMaxAttempts()">
-                    <mat-icon>remove</mat-icon>
-                </button>
-            </div>
-        </div>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.SHOWLOCKOUTFAILURES' | translate}}</span>
-            <span class="fill-space"></span>
-            <mat-slide-toggle color="primary" name="showLockOutFailures" ngDefaultControl
-                [(ngModel)]="lockoutData.showLockOutFailures">
-            </mat-slide-toggle>
-        </div>
-    </div>
-
-    <div *ngIf="policyType === PolicyComponentType?.COMPLEXITY && complexityData" class="content">
-        <mat-form-field class="description-formfield" appearance="outline">
-            <mat-label>{{ 'ORG.POLICY.DATA.DESCRIPTION' | translate }}</mat-label>
-            <input matInput name="description" ngDefaultControl [(ngModel)]="complexityData.description" required />
-        </mat-form-field>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.MINLENGTH' | translate}}</span>
-            <span class="fill-space"></span>
-            <div class="length-wrapper">
-                <button mat-icon-button (click)="decrementLength()">
-                    <mat-icon>remove</mat-icon>
-                </button>
-                <span>{{complexityData?.minLength}}</span>
-                <button mat-icon-button (click)="incrementLength()">
-                    <mat-icon>add</mat-icon>
-                </button>
-            </div>
-        </div>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.HASNUMBER' | translate}}</span>
-            <span class="fill-space"></span>
-            <mat-slide-toggle color="primary" name="hasNumber" ngDefaultControl [(ngModel)]="complexityData.hasNumber">
-            </mat-slide-toggle>
-        </div>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.HASSYMBOL' | translate}}</span>
-            <span class="fill-space"></span>
-            <mat-slide-toggle color="primary" name="hasSymbol" ngDefaultControl [(ngModel)]="complexityData.hasSymbol">
-            </mat-slide-toggle>
-        </div>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.HASLOWERCASE' | translate}}</span>
-            <span class="fill-space"></span>
-            <mat-slide-toggle color="primary" name="hasLowercase" ngDefaultControl
-                [(ngModel)]="complexityData.hasLowercase">
-            </mat-slide-toggle>
-        </div>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.HASUPPERCASE' | translate}}</span>
-            <span class="fill-space"></span>
-            <mat-slide-toggle color="primary" name="hasUppercase" ngDefaultControl
-                [(ngModel)]="complexityData.hasUppercase">
-            </mat-slide-toggle>
-        </div>
-    </div>
-
-    <div class="content" *ngIf="policyType === PolicyComponentType?.AGE && ageData">
-        <mat-form-field class="description-formfield" appearance="outline">
-            <mat-label>{{ 'ORG.POLICY.DATA.DESCRIPTION' | translate }}</mat-label>
-            <input matInput name="description" ngDefaultControl [(ngModel)]="ageData.description" required />
-        </mat-form-field>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.EXPIREWARNDAYS' | translate}}</span>
-            <span class="fill-space"></span>
-            <div class="length-wrapper">
-                <button mat-icon-button (click)="incrementExpireWarnDays()">
-                    <mat-icon>add</mat-icon>
-                </button>
-                <span>{{ageData?.expireWarnDays}}</span>
-                <button mat-icon-button (click)="decrementExpireWarnDays()">
-                    <mat-icon>remove</mat-icon>
-                </button>
-            </div>
-        </div>
-
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.MAXAGEDAYS' | translate}}</span>
-            <span class="fill-space"></span>
-            <div class="length-wrapper">
-                <button mat-icon-button (click)="incrementMaxAgeDays()">
-                    <mat-icon>add</mat-icon>
-                </button>
-                <span>{{ageData?.maxAgeDays}}</span>
-                <button mat-icon-button (click)="decrementMaxAgeDays()">
-                    <mat-icon>remove</mat-icon>
-                </button>
-            </div>
-        </div>
-    </div>
-
-    <div class="content" *ngIf="policyType === PolicyComponentType?.IAM_POLICY && iamData">
-        <mat-form-field class="description-formfield" appearance="outline">
-            <mat-label>{{ 'ORG.POLICY.DATA.DESCRIPTION' | translate }}</mat-label>
-            <input matInput name="description" ngDefaultControl [(ngModel)]="iamData.description" required />
-        </mat-form-field>
-        <div class="row">
-            <span class="left-desc">{{'ORG.POLICY.DATA.USERLOGINMUSTBEDOMAIN' | translate}}</span>
-            <span class="fill-space"></span>
-            <mat-slide-toggle color="primary" name="hasNumber" ngDefaultControl
-                [(ngModel)]="iamData.userLoginMustBeDomain">
-            </mat-slide-toggle>
-        </div>
-    </div>
-
-    <div class="btn-container">
-        <button (click)="savePolicy()" color="primary" type="submit"
-            [disabled]="(policyType === PolicyComponentType?.COMPLEXITY && !complexityData?.description) || (policyType === PolicyComponentType?.IAM_POLICY && !iamData?.description)"
-            mat-raised-button>{{ 'ACTIONS.SAVE' | translate }}</button>
-    </div>
-</app-detail-layout>
